requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
crypto==1.4.1
idna==3.6
Naked==0.1.32
pybit==5.6.2
pycryptodome==3.20.0
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.31.0
resend==0.8.0
shellescape==3.8.1
urllib3==2.2.1
websocket-client==1.7.0
websockets==12.0
python-dotenv